Use CheckBoxInput component in Match
