+++ src/components/MatchForm.js
@@ -4,24 +4,7 @@ import { Button } from 'registers-react-library';
 import { employmentBands, legalStatusBands, turnoverBands, tradingStatusBands } from '../utils/convertBands';
 import TextInput from './TextInput';
 import SelectInput from './SelectInput';
-
-const CheckBoxInput = ({ id, label, onChangeFilter, value }) => {
-  return (
-    <div className="sdc-isolation field field--checkbox field--multiplechoice">
-      <div className="field__item js-focusable-box">
-        <input onChange={onChangeFilter} value={value} className="input input--checkbox js-focusable" type="checkbox" id={id} />
-        <label className="label label--inline venus" htmlFor="checkbox">{label}</label>
-      </div>
-    </div>
-  );
-};
-
-CheckBoxInput.propTypes = {
-  id: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  onChangeFilter: PropTypes.func.isRequired,
-  value: PropTypes.bool.isRequired,
-};
+import CheckBoxInput from './CheckBoxInput';
 
 class MatchForm extends React.Component {
   // For the id of each input, we use the same name as the business-index-api input
